requirements.txt
streamlit==1.32.0
pyspark==3.5.0
kafka-python==2.0.2
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
python-dotenv==1.0.1 